PythonScript_coffee/requirements.txt
#/usr/bin/lsb_release - edit #! /usr/bin/python3.4
pyyaml
pyserial